requirements.txt
flask
requests
python-dotenv
boto3
prometheus_client


